+++ app/plugins/TwoFactorAuth/TwoFactorAuth.php
@@ -8,12 +8,14 @@
  */
 namespace Piwik\Plugins\TwoFactorAuth;
 
+use Piwik\Access;
 use Piwik\API\Request;
 use Piwik\Common;
 use Piwik\Container\StaticContainer;
 use Piwik\Exception\NoPrivilegesException;
 use Piwik\FrontController;
 use Piwik\Piwik;
+use Piwik\Plugins\Login\Controller as LoginController;
 use Piwik\Request\AuthenticationToken;
 use Piwik\Plugins\TwoFactorAuth\Dao\RecoveryCodeDao;
 use Piwik\Plugins\UsersManager\Model;
@@ -129,7 +131,7 @@ class TwoFactorAuth extends \Piwik\Plugin
             $twoFa = $this->getTwoFa();
             if ($authCode && \Piwik\Plugins\TwoFactorAuth\TwoFactorAuthentication::isUserUsingTwoFactorAuthentication($login) && $twoFa->validateAuthCode($login, $authCode)) {
                 $sessionFingerprint = new SessionFingerprint();
-                $sessionFingerprint->setTwoFactorAuthenticationVerified();
+                $sessionFingerprint->setTwoFactorAuthenticationVerified($login);
             }
         }
     }
@@ -184,6 +186,14 @@ $tokenAuth)
         if (!$this->requiresAuth($module, $action, $parameters)) {
             return;
         }
+        if ($validator->hasPendingSessionTwoFactorAuthentication() && !$validator->isCurrentUserMatchingSessionUser()) {
+            if (!Request::isRootRequestApiRequest()) {
+                $this->resetPendingTwoFactorSessionAndRequireFreshLogin($module, $action);
+            } elseif (StaticContainer::get(AuthenticationToken::class)->isSessionToken()) {
+                throw new Exception(Piwik::translate('General_YourSessionHasExpired'));
+            }
+            return;
+        }
         $twoFa = $this->getTwoFa();
         $isUsing2FA = \Piwik\Plugins\TwoFactorAuth\TwoFactorAuthentication::isUserUsingTwoFactorAuthentication(Piwik::getCurrentUserLogin());
         if ($isUsing2FA && Session::isStarted()) {
@@ -201,6 +211,13 @@ $tokenAuth)
             $module = 'TwoFactorAuth';
             $action = 'onLoginSetupTwoFactorAuth';
         }
+    }
+    private function resetPendingTwoFactorSessionAndRequireFreshLogin(&$module, &$action)
+    {
+        LoginController::clearSession();
+        Access::getInstance()->setSessionExpired(\true);
+        $module = Piwik::getLoginPluginName();
+        $action = 'login';
     }
     private function requiresAuth($module, $action, $parameters)
     {