+++ app/plugins/TwoFactorAuth/Validator.php
@@ -74,6 +74,38 @@ class Validator
             throw new Exception('not available');
         }
     }
+    public function getSessionUser()
+    {
+        $sessionFingerprint = $this->getSessionFingerPrint();
+        return $sessionFingerprint->getUser();
+    }
+    public function isCurrentUserMatchingSessionUser()
+    {
+        $sessionUser = $this->getSessionUser();
+        if (empty($sessionUser)) {
+            return \false;
+        }
+        return $sessionUser === Piwik::getCurrentUserLogin();
+    }
+    public function checkCurrentUserMatchesSessionUser()
+    {
+        if (!$this->isCurrentUserMatchingSessionUser()) {
+            throw new Exception('not available');
+        }
+    }
+    public function hasPendingSessionTwoFactorAuthentication()
+    {
+        $sessionUser = $this->getSessionUser();
+        if (empty($sessionUser)) {
+            return \false;
+        }
+        $needsTwoFactorAuthentication = \Piwik\Plugins\TwoFactorAuth\TwoFactorAuthentication::isUserUsingTwoFactorAuthentication($sessionUser) || $this->twoFa->isUserRequiredToHaveTwoFactorEnabled();
+        if (!$needsTwoFactorAuthentication) {
+            return \false;
+        }
+        $sessionFingerprint = $this->getSessionFingerPrint();
+        return !$sessionFingerprint->hasVerifiedTwoFactor();
+    }
     private function getSessionFingerPrint()
     {
         return new SessionFingerprint();